+++ src/test/java/org/stellar/sdk/responses/EffectDeserializerTest.java
@@ -2,6 +2,7 @@ package org.stellar.sdk.responses;
 
 import static org.stellar.sdk.Asset.create;
 
+import java.math.BigInteger;
 import java.util.Arrays;
 import junit.framework.TestCase;
 import org.junit.Test;
@@ -376,7 +377,8 @@ public class EffectDeserializerTest extends TestCase {
     assertEquals(effect.getType(), "claimable_balance_claimant_created");
     assertSame(effect.getPredicate().getClass(), Predicate.AbsBefore.class);
     assertEquals(
-        ((Predicate.AbsBefore) effect.getPredicate()).getTimestampSeconds(), 1234567890982222222L);
+        ((Predicate.AbsBefore) effect.getPredicate()).getTimestampSeconds(),
+        BigInteger.valueOf(1234567890982222222L));
   }
 
   @Test
@@ -419,7 +421,9 @@ public class EffectDeserializerTest extends TestCase {
         "0000000071d3336fa6b6cf81fcbeda85a503ccfabc786ab1066594716f3f9551ea4b89ca");
     assertEquals(effect.getType(), "claimable_balance_claimant_created");
     assertSame(effect.getPredicate().getClass(), Predicate.AbsBefore.class);
-    assertEquals(((Predicate.AbsBefore) effect.getPredicate()).getTimestampSeconds(), 1637479450L);
+    assertEquals(
+        ((Predicate.AbsBefore) effect.getPredicate()).getTimestampSeconds(),
+        BigInteger.valueOf(1637479450L));
   }
 
   @Test
